move LiveData ViewModel extensions to separate source file for organization
